# docs/API_SPEC.yaml
openapi: 3.1.0
...
info:
  title: Dr. Tahir LMS API
  version: 2.0.0-redesign
  description: >
    Backend for tahirelshazli.com. This version reconciles the existing API with
    the Claude Design product redesign. See docs/API_GAP_ANALYSIS.md for the
    route-by-route disposition and docs/AUTHORIZATION_MODEL.md for the rules
    every `x-roles` annotation below is shorthand for.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.tahirelshazli.com
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed. `message` is an array of constraint strings.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, expired, revoked or tampered token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >
        Authenticated but the wrong ROLE for this action. Never used for
        out-of-scope resources — those are 404.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >
        The resource does not exist, OR exists outside the caller's scope. The
        body is identical in both cases, deliberately.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The action contradicts current state (e.g. already sent).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        error: { type: string }
    Id:
      type: string
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
...
    TaskVisibility:   { type: string, enum: [published, hidden] }
...
    VisibilityState:  { type: string, enum: [published, scheduled, hidden] }
...
    StaffTaskStatus:  { type: string, enum: [open, marking, marked, closed] }
...
    SubmissionMode:   { type: string, enum: [pdf_upload, doc_link, photo_upload] }
    WorkType:         { type: string, enum: [file_upload, link, google_form] }
    AssessmentType:   { type: string, enum: [homework, assignment, quiz] }
...
    AnnotationKind:   { type: string, enum: [comment, tick, cross] }
...
    Attachment:
      type: object
      required: [url, name, audience]
      properties:
        url:       { type: string, maxLength: 2048 }
        name:      { type: string, minLength: 1, maxLength: 200 }
        mimeType:  { type: [string, 'null'], maxLength: 120 }
        sizeBytes: { type: [integer, 'null'], minimum: 0 }
        audience:  { type: string, enum: [students, staff] }

    # --- Authoring (`/staff/courses/{courseId}/assessments`, `/staff/assessments/*`)
    # Added in unit 6: these routes were `[MODIFY]` in `API_GAP_ANALYSIS.md`
    # with no entry here at all.
    AssessmentTarget:
      type: object
      required: [id, assessmentId, groupId, availableFrom, availableTo, dueAt]
      properties:
        id:            { $ref: '#/components/schemas/Id' }
        assessmentId:  { $ref: '#/components/schemas/Id' }
        groupId:       { $ref: '#/components/schemas/Id' }
        # OVERRIDES of the task's own window. Null means inherit.
        availableFrom: { type: [string, 'null'], format: date-time }
        availableTo:   { type: [string, 'null'], format: date-time }
        dueAt:         { type: [string, 'null'], format: date-time }
    AssessmentTargetInput:
      type: object
      required: [groupId]
      properties:
        groupId:       { $ref: '#/components/schemas/Id' }
        availableFrom: { type: string, format: date-time }
        availableTo:   { type: string, format: date-time }
        dueAt:         { type: string, format: date-time }

    # A task as the authoring screens see it: every stored field plus its full
    # audience. `visibility`, `markerId`, `allowResubmission`, `submissionModes`,
    # `draftId` and `attachments` are unit 6 (migration 018).
    AuthoredAssessment:
      type: object
      required: [id, courseId, lessonId, title, description, instructions, type, topics,
                 availableFrom, availableTo, dueAt, maxScore, allowedFileTypes,
                 maxFileSizeBytes, workType, externalUrl, visibility, markerId,
                 allowResubmission, submissionModes, draftId, attachments, createdAt, targets]
      properties:
        id:          { $ref: '#/components/schemas/Id' }
        courseId:    { $ref: '#/components/schemas/Id' }
        lessonId:    { type: [string, 'null'] }
        title:       { type: string }
        description: { type: string }
        instructions: { type: string }
        type:        { $ref: '#/components/schemas/AssessmentType' }
        topics:      { type: array, items: { type: string } }
        availableFrom: { type: string, format: date-time }
        availableTo:   { type: string, format: date-time }
        dueAt:         { type: string, format: date-time }
        maxScore:      { type: integer }
        allowedFileTypes: { type: array, items: { type: string } }
        maxFileSizeBytes: { type: integer }
        workType:    { $ref: '#/components/schemas/WorkType' }
        externalUrl: { type: [string, 'null'] }
        visibility:  { $ref: '#/components/schemas/TaskVisibility' }
        markerId:    { type: [string, 'null'] }
        # `true` (the default) is today's rule: resubmission until window END,
        # not `dueAt` (`D-31`). `false` makes a second submission a 409.
        allowResubmission: { type: boolean }
        submissionModes: { type: array, items: { $ref: '#/components/schemas/SubmissionMode' } }
        # Provenance only - the content was COPIED from the draft, never linked.
        # Null once the draft is deleted.
        draftId:     { type: [string, 'null'] }
        attachments: { type: array, items: { $ref: '#/components/schemas/Attachment' } }
        createdAt:   { type: string, format: date-time }
        targets:     { type: array, items: { $ref: '#/components/schemas/AssessmentTarget' } }

    # A target on the staff task list, with its group's display name.
    StaffTaskTarget:
      allOf:
        - { $ref: '#/components/schemas/AssessmentTarget' }
        - type: object
          required: [groupName]
          properties:
            groupName: { type: string }

    # One row of `GET /staff/tasks`: every `AuthoredAssessment` field, with
    # `targets` narrowed to the groups the caller reaches.
    StaffTask:
      allOf:
        - { $ref: '#/components/schemas/AuthoredAssessment' }
        - type: object
          required: [targets, visibilityState, status, markerName, markerDrift]
          properties:
            targets: { type: array, items: { $ref: '#/components/schemas/StaffTaskTarget' } }
            visibilityState: { $ref: '#/components/schemas/VisibilityState' }
            # `D-30`/`D-34`/`D-35`. Always one of the four.
            status: { $ref: '#/components/schemas/StaffTaskStatus' }
            markerName: { type: [string, 'null'] }
            # `D-32`: the named marker no longer qualifies for the task's WHOLE
            # audience. Displayed, never silently repaired. A fact about the
            # task, identical for every viewer. False when no marker is named.
            markerDrift: { type: boolean }

    # `POST /staff/courses/{courseId}/assessments`. The body is authoritative:
    # `draftId` records provenance and bumps the draft's `usedCount`, and the
    # server merges nothing from the draft (assumption A-2).
    AssessmentWrite:
      type: object
      required: [title, type, availableFrom, availableTo, dueAt, maxScore,
                 allowedFileTypes, maxFileSizeBytes, targets]
      properties:
        title:        { type: string, minLength: 1, maxLength: 200 }
        description:  { type: string, maxLength: 1000 }
        instructions: { type: string, maxLength: 5000 }
        type:         { $ref: '#/components/schemas/AssessmentType' }
        topics:       { type: array, maxItems: 20, items: { type: string, maxLength: 80 } }
        lessonId:     { $ref: '#/components/schemas/Id' }
        availableFrom: { type: string, format: date-time }
        availableTo:   { type: string, format: date-time }
        dueAt:         { type: string, format: date-time }
        maxScore:      { type: integer, minimum: 1, maximum: 1000 }
        allowedFileTypes: { type: array, minItems: 1, maxItems: 20, items: { type: string } }
        maxFileSizeBytes: { type: integer, minimum: 1, maximum: 104857600 }
        workType:    { $ref: '#/components/schemas/WorkType' }
        externalUrl: { type: string, maxLength: 2048 }
        googleForm:  { type: string, minLength: 8, maxLength: 512 }
        targets:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: '#/components/schemas/AssessmentTargetInput' }
        draftId:     { $ref: '#/components/schemas/Id' }
        attachments:
          type: array
          maxItems: 10
          items: { $ref: '#/components/schemas/Attachment' }
        allowResubmission: { type: boolean, default: true }
        visibility: { $ref: '#/components/schemas/TaskVisibility' }
        # `D-32`. Null/omitted is "whoever opens it first". An assistant sending a
        # non-null value is a 403. The named user must be the teacher, an admin,
        # or an ACTIVE assistant who reaches every targeted group - else 400.
        markerId: { type: [string, 'null'] }
        # `D-31`. Each at most once. Omitted is `[]` ("not stated").
        submissionModes:
          type: array
          uniqueItems: true
          maxItems: 3
          items: { $ref: '#/components/schemas/SubmissionMode' }

    # `PATCH /staff/assessments/{assessmentId}`. Every field optional and never
    # null. `type`, `courseId` and `draftId` are not patchable.
    AssessmentPatch:
      type: object
      properties:
        title:        { type: string, minLength: 1, maxLength: 200 }
        description:  { type: string, maxLength: 1000 }
        instructions: { type: string, maxLength: 5000 }
        topics:       { type: array, maxItems: 20, items: { type: string, maxLength: 80 } }
        availableFrom: { type: string, format: date-time }
        availableTo:   { type: string, format: date-time }
        dueAt:         { type: string, format: date-time }
        maxScore:      { type: integer, minimum: 1, maximum: 1000 }
        allowedFileTypes: { type: array, minItems: 1, maxItems: 20, items: { type: string } }
        maxFileSizeBytes: { type: integer, minimum: 1, maximum: 104857600 }
        workType:    { $ref: '#/components/schemas/WorkType' }
        externalUrl: { type: string, maxLength: 2048 }
        googleForm:  { type: string, minLength: 8, maxLength: 512 }
        # Replaces the whole list; `[]` clears it.
        attachments:
          type: array
          maxItems: 10
          items: { $ref: '#/components/schemas/Attachment' }
        allowResubmission: { type: boolean }
        # `hidden` on a task with any submission is a 409 (`D-28`).
        visibility: { $ref: '#/components/schemas/TaskVisibility' }
        # `D-32`. Teacher/admin only: an assistant setting OR clearing a marker
        # is a 403. Checked against the task's current audience (400); re-aiming
        # the task later does not clear it.
        markerId: { type: [string, 'null'] }
        # `D-31`. Replaces the whole set.
        submissionModes:
          type: array
          uniqueItems: true
          maxItems: 3
          items: { $ref: '#/components/schemas/SubmissionMode' }

    # --- Marking -----------------------------------------------------------
    Annotation:
      type: object
      required: [id, submissionId, page, xPercent, yPercent, kind, text]
      properties:
        id:           { $ref: '#/components/schemas/Id' }
        submissionId: { $ref: '#/components/schemas/Id' }
        page:         { type: integer, minimum: 1 }
        xPercent:     { type: number, minimum: 0, maximum: 100 }
        yPercent:     { type: number, minimum: 0, maximum: 100 }
        kind:         { $ref: '#/components/schemas/AnnotationKind' }
        text:         { type: string, maxLength: 2000 }
        createdBy:    { $ref: '#/components/schemas/Id' }
        createdAt:    { type: string, format: date-time }
    AnnotationWrite:
      type: object
      required: [page, xPercent, yPercent, kind, text]
      properties:
        page:     { type: integer, minimum: 1 }
        xPercent: { type: number, minimum: 0, maximum: 100 }
        yPercent: { type: number, minimum: 0, maximum: 100 }
        kind:     { $ref: '#/components/schemas/AnnotationKind' }
        text:     { type: string, maxLength: 2000 }

    # --- Sessions ----------------------------------------------------------
...
paths:
...
  /staff/courses/{courseId}/assessments:
    parameters:
      - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    get:
      tags: [tasks]
      summary: Every task on one course, with its full audience
      description: >
        Course-grained (`AUTH-6` remainder): one held cohort reaches every task on
        the course. The group-grain list is `GET /staff/tasks`.
      x-roles: [assistant, teacher, admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AuthoredAssessment' } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [tasks]
      x-roles: [assistant, teacher, admin]
      description: >
        A `draftId` that is missing, on another course, or unreachable is one
        identical 404 (`Task draft not found`); authoring from a draft adds one to
        its `usedCount` in the same transaction. `D-33`: a target group the caller
        does not hold is a 404 with the same message a group not on the course
        gets (`Group <id> is not enrolled in this course`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssessmentWrite' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthoredAssessment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
      x-audit: assessment.created
  /staff/assessments/{assessmentId}:
    parameters:
      - { name: assessmentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    patch:
      tags: [tasks]
      x-roles: [assistant, teacher, admin]
      description: >
        403 when an assistant sets or clears `markerId` (`D-32`) - the task is on
        their screen, so the anti-enumeration 404 does not apply.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssessmentPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthoredAssessment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`visibility: hidden` on a task that has any submission (`D-28`) or any synced external result (`D-36`).'
      x-audit: assessment.updated
    delete:
      tags: [tasks]
      x-roles: [assistant, teacher, admin]
      description: >
        Refused with 400 once anything has been submitted (pre-existing), and with
        409 once any external result has synced (`D-36`) - deleting would cascade
        students' form responses away.
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
      x-audit: assessment.deleted
  /staff/assessments/{assessmentId}/targets:
    parameters:
      - { name: assessmentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    post:
      tags: [tasks]
      summary: Replace the whole audience (a set, not a diff)
      description: >
        `D-33`. A scoped assistant re-aiming a task whose CURRENT audience includes
        a group they cannot reach is a 403 - refused rather than silently dropping
        groups they cannot see (the task is on their screen). Adding a group they
        do not hold is a 404 whose message is byte-identical to
        `Group <id> is not enrolled in this course` for that id. The teacher, an
        admin and an `all_groups` assistant are unaffected.
      x-roles: [assistant, teacher, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targets]
              properties:
                targets:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: { $ref: '#/components/schemas/AssessmentTargetInput' }
      responses:
        '201':
          description: The task with its new audience
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthoredAssessment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
      x-audit: assessment.targeted
...
  /courses/{courseId}/assessments:
    get:
      tags: [tasks]
      summary: The student's task list for one course
      description: >
        A `hidden` task (`D-28`) is absent. A `published` task with a future window
        is present as `locked`, with its `availableFrom` - that is what "scheduled"
        means to a student.
      x-roles: [student]
      parameters:
        - { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
  /assessments/{assessmentId}:
    get:
      tags: [tasks]
      summary: One task as the student sees it
      description: >
        Carries `attachments` - ONLY those whose `audience` is `students`
        (`D-29`), each without the `audience` field. A `hidden` task (`D-28`) is a
        404 whose body is identical to a nonexistent id. `canSubmit` is false once
        a one-shot task (`allowResubmission: false`) has a submission.
      x-roles: [student]
      parameters:
        - { name: assessmentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /staff/tasks:
    get:
      tags: [tasks]
      summary: Every task in scope, across courses - group-grain from birth
      description: >
        Backs the console's Tasks list (`TASK-6`). A task is listed only through a
        target the caller reaches (`StaffScopeService.reachableGroupIds`), and each
        row's `targets` are narrowed to those groups IN THE QUERY - an assistant
        never receives an unheld group's id or name. Every filter narrows and none
        addresses a resource: an unheld or unknown `groupId` and an unreachable
        `courseId` all answer `200 []`. `search` is a case-insensitive literal
        title substring (`%` and `_` are not wildcards). Ordered `dueAt` desc, id.
        Not paged (two courses, ~20 tasks each). No per-row submission counts
        (`D-30`).
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: courseId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { name: groupId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - name: status
          in: query
          description: >
            `D-30`/`D-34`/`D-35`. Derived on the server, never stored. `open` =
            now <= the LATEST of the targeted groups' own due dates (override or
            the task's); after that, `marking` = any ungraded submission, `marked`
            = every submission graded, `closed` = no submission at all (link and
            Google Form work included). Judged over the task's WHOLE audience and
            every submission, so identical for every viewer. Never null.
          schema: { $ref: '#/components/schemas/StaffTaskStatus' }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StaffTask' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # --- Marking ------------------------------------------------------------
...
  /staff/submissions/{submissionId}/annotations:
    parameters:
      - { name: submissionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    get:
      tags: [marking]
      x-roles: [assistant, teacher, admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Annotation' } }
    post:
      tags: [marking]
      x-roles: [assistant, teacher, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnotationWrite' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Annotation' }
      x-audit: submission.annotated
...
  /staff/submissions/{submissionId}/return:
    post:
      tags: [marking]
      summary: Return the marked paper to the student
      description: >
        Distinct from saving a mark. Saving records the score and annotations;
        returning is what makes them visible. The original file is never
        replaced — the annotated copy sits alongside it.
      x-roles: [assistant, teacher, admin]
      parameters:
        - { name: submissionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                includeInReport: { type: boolean, default: true }
      responses:
        '200': { description: Returned }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
      x-audit: submission.returned

  # --- Mark book ----------------------------------------------------------
...